Move backend api tests
