+++ packages/backend/src/api/routers/FrontendRouter.test.ts
@@ -8,21 +8,21 @@ import {
 } from '@explorer/types'
 import { SuperAgentTest } from 'supertest'
 
-import { ControllerSuccessResult } from '../../../src/api/controllers/ControllerResult'
-import { ForcedTradeOfferController } from '../../../src/api/controllers/ForcedTradeOfferController'
-import { ForcedTransactionController } from '../../../src/api/controllers/ForcedTransactionController'
-import { HomeController } from '../../../src/api/controllers/HomeController'
-import { PositionController } from '../../../src/api/controllers/PositionController'
-import { SearchController } from '../../../src/api/controllers/SearchController'
-import { StateUpdateController } from '../../../src/api/controllers/StateUpdateController'
-import { createFrontendRouter } from '../../../src/api/routers/FrontendRouter'
-import { PositionRepository } from '../../../src/peripherals/database/PositionRepository'
-import { StateUpdateRepository } from '../../../src/peripherals/database/StateUpdateRepository'
-import { UserRegistrationEventRepository } from '../../../src/peripherals/database/UserRegistrationEventRepository'
-import { Logger } from '../../../src/tools/Logger'
-import { mock } from '../../mock'
-import { setupDatabaseTestSuite } from '../../peripherals/database/shared/setup'
-import { createTestApiServer } from '../TestApiServer'
+import { mock } from '../../../test/mock'
+import { setupDatabaseTestSuite } from '../../../test/peripherals/database/shared/setup'
+import { PositionRepository } from '../../peripherals/database/PositionRepository'
+import { StateUpdateRepository } from '../../peripherals/database/StateUpdateRepository'
+import { UserRegistrationEventRepository } from '../../peripherals/database/UserRegistrationEventRepository'
+import { createTestApiServer } from '../../test/TestApiServer'
+import { Logger } from '../../tools/Logger'
+import { ControllerSuccessResult } from '../controllers/ControllerResult'
+import { ForcedTradeOfferController } from '../controllers/ForcedTradeOfferController'
+import { ForcedTransactionController } from '../controllers/ForcedTransactionController'
+import { HomeController } from '../controllers/HomeController'
+import { PositionController } from '../controllers/PositionController'
+import { SearchController } from '../controllers/SearchController'
+import { StateUpdateController } from '../controllers/StateUpdateController'
+import { createFrontendRouter } from './FrontendRouter'
 
 const TEST_PAGE = '<!DOCTYPE html><p>test page</p>'
 const SUCCESS_RESULT: ControllerSuccessResult = {
