+++ packages/backend/src/test/TestApiServer.ts
@@ -2,8 +2,8 @@ import Router from '@koa/router'
 import { Middleware } from 'koa'
 import { agent } from 'supertest'
 
-import { ApiServer } from '../../src/api/ApiServer'
-import { Logger } from '../../src/tools/Logger'
+import { ApiServer } from '../api/ApiServer'
+import { Logger } from '../tools/Logger'
 
 export function createTestApiServer(
   routers: Router[],
